Cutover to storing each proposal in GCS instead.
The MySQL hosting on GCP was super expensive with default settings; few
dollars a day for it to just sit idle! I realized a much simpler
architecture -- just stick each proposal as a blob in cloud storage,
indexed by checksum. This is really a glorified key/value store. If we
want to do some kind of proposal browsing and moderation, it can happen
with something else tacked on; we don't need a full SQL DB to do that
yet.